requirements.txt
# Core runtime
numpy>=1.23,<3.0

# Audio I/O
sounddevice>=0.4.6,<0.5

# TTS
piper-tts>=1.2,<2.0

# ASR
faster-whisper>=1.0,<2.0

# LLM + LangChain
langchain>=0.2.12,<0.3
langchain-community>=0.2.12,<0.3
llama-cpp-python>=0.2.90,<0.3